GenericAutoContainer: added advanced settings
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@32134 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/decrypter/GenericAutoContainer.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 
 import jd.PluginWrapper;
 import jd.controlling.ProgressController;
+import jd.controlling.linkcrawler.LinkCrawlerConfig;
 import jd.http.URLConnectionAdapter;
 import jd.parser.Regex;
 import jd.plugins.CryptedLink;
@@ -15,6 +16,7 @@ import jd.plugins.Plugin;
 import jd.plugins.PluginForDecrypt;
 import jd.utils.JDUtilities;
 
+import org.appwork.storage.config.JsonConfig;
 import org.appwork.utils.StringUtils;
 
 @DecrypterPlugin(revision = "$Revision$", interfaceVersion = 2, names = { "genericautocontainer" }, urls = { "https?://[\\w\\.:\\-@]*/.*\\.(dlc|ccf|rsdf)$" }, flags = { 0 })
@@ -28,37 +30,41 @@ public class GenericAutoContainer extends PluginForDecrypt {
     public ArrayList<DownloadLink> decryptIt(CryptedLink parameter, ProgressController progress) throws Exception {
         final ArrayList<DownloadLink> ret = new ArrayList<DownloadLink>();
         final String url = parameter.getCryptedUrl();
-        final String type = new Regex(url, "(dlc|rsdf|ccf)$").getMatch(0);
-        URLConnectionAdapter con = null;
-        File containerTemp = null;
-        try {
-            con = br.openGetConnection(url);
-            if (con.isOK()) {
-                boolean seemsValidContainer = StringUtils.containsIgnoreCase(con.getContentType(), type);
-                seemsValidContainer = seemsValidContainer | (con.isContentDisposition() && StringUtils.containsIgnoreCase(Plugin.getFileNameFromHeader(con), type));
-                if (seemsValidContainer) {
-                    containerTemp = org.appwork.utils.Application.getResource("tmp/autocontainer/" + System.nanoTime() + "." + type);
-                    br.downloadConnection(containerTemp, con);
-                    if (containerTemp.exists() && containerTemp.length() > 100) {
-                        ret.addAll(JDUtilities.getController().getContainerLinks(containerTemp));
+        if (!JsonConfig.create(LinkCrawlerConfig.class).isAutoImportContainer()) {
+            ret.add(createDownloadlink(url));
+        } else {
+            final String type = new Regex(url, "(dlc|rsdf|ccf)$").getMatch(0);
+            URLConnectionAdapter con = null;
+            File containerTemp = null;
+            try {
+                con = br.openGetConnection(url);
+                if (con.isOK()) {
+                    boolean seemsValidContainer = StringUtils.containsIgnoreCase(con.getContentType(), type);
+                    seemsValidContainer = seemsValidContainer | (con.isContentDisposition() && StringUtils.containsIgnoreCase(Plugin.getFileNameFromHeader(con), type));
+                    if (seemsValidContainer) {
+                        containerTemp = org.appwork.utils.Application.getResource("tmp/autocontainer/" + System.nanoTime() + "." + type);
+                        br.downloadConnection(containerTemp, con);
+                        if (containerTemp.exists() && containerTemp.length() > 100) {
+                            ret.addAll(JDUtilities.getController().getContainerLinks(containerTemp));
+                        }
                     }
                 }
+            } catch (final IOException e) {
+                logger.log(e);
+                if (ret.size() == 0) {
+                    ret.add(createDownloadlink(url));
+                }
+            } finally {
+                if (containerTemp != null && containerTemp.exists()) {
+                    containerTemp.delete();
+                }
+                if (con != null) {
+                    con.disconnect();
+                }
             }
-        } catch (final IOException e) {
-            logger.log(e);
-            if (ret.size() == 0) {
+            if (containerTemp != null && ret.size() == 0) {
                 ret.add(createDownloadlink(url));
             }
-        } finally {
-            if (containerTemp != null && containerTemp.exists()) {
-                containerTemp.delete();
-            }
-            if (con != null) {
-                con.disconnect();
-            }
-        }
-        if (containerTemp != null && ret.size() == 0) {
-            ret.add(createDownloadlink(url));
         }
         return ret;
     }
